fix ./configure on macOS
Without this change, ./configure thinks stdlib.h is not available. I am
slightly surprised ./configure (presumably) worked on other systems
without this change since POSIX specifies that exit(3) is declared in
stdlib.h. xplot was written in the late 1980s -- perhaps exit(3) was
declared elsewhere then?

Speaking of the age of xplot, it was fun seeing references to Ultrix.
That _really_ dates this software.